Implement stealth modus.
In stealth modus, the connection will ensure that it will not send more
packets, nor more data, than what it has received from the remote end.
Stealth is disengaged once the negotiation completes
(net2_conn_negotiator is responsible for disengaging stealth).

The stealth mode is to avoid being a juicy target for a deflector style
DOS attack, where this connection is used as a method to grow the size
of such an attack (for example by sending small packets repeatedly,
which would cause large packets to be sent to the actual target).