Allow cloning of xchange context.
A cloned xchange has the same parameters as the original context had
right after calling net2_xchangectx_prepare().
This can be used to reduce the overhead that generating a new exchange
contexts incurs.

Note that the same effect could be achieved by importing the buffer
generated by a prepare() call.