Implement sensitive buffers.
When a buffer is marked as sensitive, all segments it contains at that
moment will be secure_zeroed before free.
Any new data added to the buffer is not automatically considered secure.

References will never be secure_zeroed, this is the responsibility of
its release/free implementation (the reasoning being that the free
function of a reference might be something other than an actual call
to free).