Detachable threads, report net2_init() failure.
Threads under windows require something to store their thread data
structure in a place we can find it.  Since win32 TLS is only a
DWORD (32 bit), we'd require at least 2 pointers to get those stored
correctly.  It's complicated and may cause pressure on the limited
TLS storage area.

Since the tree is only required to be able to detach threads,
simply store the threads in a tree.