First version of datapipe.
A datapipe is a thread-safe queue, which has one or more producers and
one or more consumers.  Producers and consumers can be manual
using the net2_datapipe_in (input side of the queue, aka producer side)
and net2_datapipe_out (output side of the queue, aka consumer side).
In addition, events can be attached to the queue, making the input
and output side invoke callbacks.

For event callbacks, an optional workq can be specified, in which case
the event callback will synchronize to this workq while running the
callback.  If no workq is specified, no such synchronization takes
place.